src/os/create-process-unix.lisp: (process-wait): kludge for lispworks
LispWorks sets up a waitpid(0) handler for all its subprocesses and
gets the signal before iolib and when iolib calls process-wait it only
gets an ECHILD as the process has already been reaped.

Try to handle that case. Return the process-pid which we already have
for the process which was passed to us and return a status of -10
Regardless of what the actual status was. Only LispWorks knows that.